Change entrypoint to run rspec and remove unneeded containers
We don't need the seed container (seeding it handled by rspec's
helpers during testing) and we don't shouldn't need sidekiq (expecting
services are test time to be running feels like a problem, we might
actually have this).

This probably won't work for front-end builds (where we expect to be
in production mode for some integration/e2e tests) but it's a start -
right now this gets me back to the failures linked to carrierwave.