Clear the memoized class variable before and after finding release
We do something like this in the ForemInstance unit tests for the
underlying functionality - I would rather we didn't need to know this
indirect collaborator side-effect (the setup conditions need to cross
an abstraction boundary and know that the ForemInstance has a
conditional based on the release footprint value - and also need to
know that it's memoizing that value and that they need to clear it).

If reaching into the internals of this class with
`instance_variable_set` is required as often as we do it in the tests,
there's probably some improvement possible.